backend/pyproject.toml
[project]
name = "warmai-backend"
version = "0.2.0"
description = "WarmAI backend — email warmup automation API + scheduler"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic[email]>=2.9.0",
    "pydantic-settings>=2.6.0",
    "beanie>=1.27.0,<2.0",
    "motor>=3.6.0,<4.0",
    "pymongo>=4.6,<5.0",
    "apscheduler>=3.10.4",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "python-multipart>=0.0.12",
    "pandas>=2.2.0",
    "chardet>=5.2.0",
    "structlog>=24.4.0",
    "httpx>=0.27.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mongomock-motor>=0.0.34",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "RUF"]
ignore = ["E501"]
